feat: env-configurable MarketDataCore, theme system, and client tracking
Core:
- MarketDataCoreEngine now reads SENTINEL_MDC_{HOST,PORT,TARGET,USE_JWT} env vars
- Add subscription logging with JWT redaction for debugging
- SentinelStreamServer emits clientSubscribed/clientUnsubscribed signals
- Change heatmap column format from u8 to u16

GUI:
- Introduce ThemeBridge + FontManager for centralized theming
- Add TopToolbar, WatchlistDock, and ChartToolRail widgets
- Expand DarkTheme stylesheet (QTreeView, improved menus)
- Add Qt6 Svg component and icon resources

+++ libs/gui/render/HeatmapLabelRenderer.hpp
@@ -21,7 +21,7 @@ public:
     static void buildLabelQuads(const HeatmapStreamState::Snapshot& snapshot,
                                 const GlyphAtlas& atlas,
                                 const std::vector<uint16_t>& liquidityRing,
-                                const std::vector<uint8_t>& intensityRing,
+                                const std::vector<uint16_t>& intensityRing,
                                 const std::vector<double>& liquidityScales,
                                 const QRectF& srcRect,
                                 const QRectF& drawRect,
